feat: move parse user page link to lib/user
