+++ components/deals/deal-thumbnail-preview.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import { ChatBubbleOvalLeftIcon } from '@heroicons/react/24/outline';
 import { UserImageResponse } from '@/generated/graphql';
 import { Deal, DealStatus } from '@/lib/deal/deal.types';
-import { parseDealDetailLink } from '@/lib/link/parse-user-page.link';
+import { parseDealDetailLink } from '@/lib/user/parse-user-page.link';
 import Thumbnail from '../base/thumbnail';
 import DealMenu from './deal-menu';
 import DealAddons from './deal-addons';
